tmp working with many template args
