+++ core/src/MeanData.h
@@ -1,5 +1,5 @@
-#ifndef MAQ_HULLDATA_H
-#define MAQ_HULLDATA_H
+#ifndef MAQ_MEANDATA_H
+#define MAQ_MEANDATA_H
 
 #include <cstddef>
 #include <vector>
@@ -8,51 +8,10 @@
 
 namespace maq {
 
-class HullData {
+template <Storage storage, SampleWeights sample_weights, TieBreaker tie_breaker>
+class MeanData {
 public:
-  virtual double get_reward(size_t row, size_t col) const = 0;
-
-  virtual double get_cost(size_t row, size_t col) const = 0;
-
-  virtual size_t get_num_rows() const = 0;
-
-  virtual size_t get_num_cols() const = 0;
-
-  virtual double get_reward_scores(size_t row, size_t col) const = 0;
-};
-
-class CompleteData : public HullData {
-public:
-  CompleteData(const Data& data) :
-    data(data) {}
-
-  double get_reward(size_t row, size_t col) const {
-    return data.get_reward(row, col);
-  }
-
-  double get_cost(size_t row, size_t col) const {
-    return data.get_cost(row, col);
-  }
-
-  size_t get_num_rows() const {
-    return data.num_rows;
-  }
-
-  size_t get_num_cols() const {
-    return data.num_cols;
-  }
-
-  double get_reward_scores(size_t row, size_t col) const {
-    return data.get_reward_scores(row, col);
-  }
-
-private:
-  const Data& data;
-};
-
-class MeanData : public HullData {
-public:
-  MeanData(const Data& data, const std::vector<size_t>& samples) {
+  MeanData(const Data<storage, sample_weights, tie_breaker>& data, const std::vector<size_t>& samples) {
     std::vector<double> reward(data.num_cols);
     std::vector<double> reward_scores(data.num_cols);
     std::vector<double> cost(data.num_cols);
@@ -98,4 +57,4 @@ private:
 
 } // namespace maq
 
-#endif // MAQ_HULLDATA_H
+#endif // MAQ_MEANDATA_H
